productinventory

+++ src/main/java/com/nextech/erp/model/Client.java
@@ -43,7 +43,7 @@ public class Client implements Serializable {
 	private String contactpersonname;
 
 	@Column(name="created_by")
-	private int createdBy;
+	private long createdBy;
 
 	@Column(name="created_date")
 	private Timestamp createdDate;
@@ -59,7 +59,7 @@ public class Client implements Serializable {
 	private boolean isactive;
 
 	@Column(name="updated_by")
-	private int updatedBy;
+	private long updatedBy;
 
 	@Column(name="updated_date")
 	private Timestamp updatedDate;
@@ -115,11 +115,11 @@ public class Client implements Serializable {
 		this.contactpersonname = contactpersonname;
 	}
 
-	public int getCreatedBy() {
+	public long getCreatedBy() {
 		return this.createdBy;
 	}
 
-	public void setCreatedBy(int createdBy) {
+	public void setCreatedBy(long createdBy) {
 		this.createdBy = createdBy;
 	}
 
@@ -155,11 +155,11 @@ public class Client implements Serializable {
 		this.isactive = isactive;
 	}
 
-	public int getUpdatedBy() {
+	public long getUpdatedBy() {
 		return this.updatedBy;
 	}
 
-	public void setUpdatedBy(int updatedBy) {
+	public void setUpdatedBy(long updatedBy) {
 		this.updatedBy = updatedBy;
 	}
